requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
openai==1.3.7
python-dotenv==1.0.0
gunicorn==21.2.0

